requirements.txt
# requirements.txt
fastapi
uvicorn
pymongo
mysql-connector-python
kafka-python
python-dotenv
websockets
requests
paho-mqtt
